feat: Complete ECCVM recursive verifier (#6720)
Enable IPA recursive verification and add tests to complete the ECCVM
recursive verifier.

---------

+++ barretenberg/cpp/src/barretenberg/stdlib/primitives/bigfield/bigfield.hpp
@@ -253,7 +253,7 @@ template <typename Builder, typename T> class bigfield {
      * @return bigfield
      */
     bigfield pow(const field_t<Builder>& exponent) const;
-    bigfield pow(const uint32_t exponent) const;
+    bigfield pow(const size_t exponent) const;
     bigfield madd(const bigfield& to_mul, const std::vector<bigfield>& to_add) const;
 
     static void perform_reductions_for_mult_madd(std::vector<bigfield>& mul_left,
